Refactor: 채팅 관련 코드 수정
