+++ src/pages/Chatting.tsx
@@ -4,16 +4,22 @@ import { useState, useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { VscMenu, VscArrowLeft } from 'react-icons/vsc';
 import { Client } from '@stomp/stompjs';
+import { currentTime, TimeFormatEnum } from '@src/util/CurrentTime';
 
-type Props = {};
+export interface ChatListType {
+  person: string;
+  time: string;
+  message: string;
+}
 
 // 접속id와 다를경우 접근 불가
-const Chatting = (props: Props) => {
+const Chatting = () => {
   const roomId = useParams();
   const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
   const [recievedMessage, setRecievedMessage] = useState<string>('');
   const [myMessage, setMyMessage] = useState<string>('');
-  const [chatList, setChatList] = useState<string[]>([]);
+  const [chatList, setChatList] = useState<Array<ChatListType>>([]);
+  const time = currentTime(1000, TimeFormatEnum.HHmm);
   const [chatDate, setChatDate] = useState();
   // 서버에서 받아온 이전 채팅기록/오늘날짜 비교 -> 이전기록이 오늘일 경우 라벨 x
 
@@ -36,7 +42,7 @@ const Chatting = (props: Props) => {
         message.body.length >= 1 && setRecievedMessage(message.body);
       });
       client.publish({
-        destination: `/pub/chat/room/4`,
+        destination: `/pub/chat/send`,
         body: myMessage,
       });
     },
@@ -55,12 +61,13 @@ const Chatting = (props: Props) => {
   }, []);
 
   useEffect(() => {
-    recievedMessage.length >= 1 && setChatList((prev) => [...prev, recievedMessage]);
-  }, [recievedMessage]);
-
-  useEffect(() => {
-    myMessage.length >= 1 && setChatList((prev) => [...prev, myMessage]);
+    myMessage.length >= 1 &&
+      setChatList((prev) => [...prev, { person: 'me', time: time, message: myMessage }]);
   }, [myMessage]);
+  useEffect(() => {
+    recievedMessage.length >= 1 &&
+      setChatList((prev) => [...prev, { person: 'other', time: time, message: recievedMessage }]);
+  }, [recievedMessage]);
 
   const reservation = () => {
     //거래예약 api
@@ -116,7 +123,7 @@ const Chatting = (props: Props) => {
           <span className='mr-2'>3월 4일 토요일</span>/<span className='ml-2'>4만원</span>
         </div>
       </section>
-      <ChatBox myMessage={myMessage} recievedMessage={recievedMessage} />
+      <ChatBox chatList={chatList} />
       <SendMessage setMyMessage={setMyMessage} />
     </main>
   );
